requirements.txt
# Pinned for Python 3.11
# Top-level runtime dependencies inferred from the current codebase.

firecrawl-py==2.16.5
cryptography==46.0.3
email-validator==2.3.0
fastapi==0.124.0
groq==1.1.2
httpx==0.28.1
langgraph==1.1.6
openai==2.31.0
passlib[bcrypt]==1.7.4
psycopg[binary]==3.3.2
pydantic==2.13.0
PyJWT==2.10.1
PyMuPDF==1.27.2.2
python-dotenv==1.2.2
requests==2.33.1
sqlalchemy==2.0.45
tavily-python==0.7.23
trafilatura==2.0.0
uvicorn[standard]==0.38.0
